Moving schema extensions into the correct place
